Show progress when moving library to a new location

+++ src/calibre/library/database2.py
@@ -908,11 +908,21 @@ class LibraryDatabase2(LibraryDatabase):
         self.data.books_added([id], self.conn)
         self.notify('add', [id])
         
-    def move_library_to(self, newloc):
+    def move_library_to(self, newloc, progress=None):
+        header = _(u'<p>Copying books to %s<br><center>')%newloc
+        books = self.conn.execute('SELECT id, path, title FROM books').fetchall()
+        if progress is not None:
+            progress.setValue(0)
+            progress.setLabelText(header)
+            QCoreApplication.processEvents()
+            progress.setAutoReset(False)
+            progress.setRange(0, len(books))
         if not os.path.exists(newloc):
             os.makedirs(newloc)
         old_dirs = set([])
-        for book in self.conn.execute('SELECT id, path FROM books').fetchall():
+        for i, book in enumerate(books):
+            if progress is not None:
+                progress.setLabelText(header+_(u'Copying <b>%s</b>')%book[2])
             path = book[1]
             if not path:
                 continue
@@ -921,8 +931,11 @@ class LibraryDatabase2(LibraryDatabase):
             tdir = os.path.join(newloc, dir)
             if os.path.exists(tdir):
                 shutil.rmtree(tdir)
-            shutil.copytree(srcdir, tdir)
+            if os.path.exists(srcdir):
+                shutil.copytree(srcdir, tdir)
             old_dirs.add(srcdir)
+            if progress is not None:
+                progress.setValue(i+1)
         
         dbpath = os.path.join(newloc, os.path.basename(self.dbpath))
         shutil.copyfile(self.dbpath, dbpath)
@@ -936,6 +949,9 @@ class LibraryDatabase2(LibraryDatabase):
                 shutil.rmtree(dir)
         except:
             pass
+        if progress is not None:
+            progress.reset()
+            progress.hide()
             
     
     def migrate_old(self, db, progress):
